changed address geocoder to a class
