+++ lib/address_geocoder/request.rb
@@ -1,23 +1,25 @@
 require 'httparty'
 
-class Request # :nodoc:
-  attr_accessor :result
+class AddressGeocoder
+  class Request # :nodoc:
+    attr_accessor :result
 
-  def initialize(url)
-    attempts = 0
-    begin
-      @result = HTTParty.get(url)
-    rescue
-      sleep(0.5)
-      attempts += 1
-      retry if attempts <= 5
-      raise SystemCallError, 'Could not connect to GoogleAPI'
+    def initialize(url)
+      attempts = 0
+      begin
+        @result = HTTParty.get(url)
+      rescue
+        sleep(0.5)
+        attempts += 1
+        retry if attempts <= 5
+        raise SystemCallError, 'Could not connect to GoogleAPI'
+      end
     end
-  end
 
-  def success?
-    return false unless @result['status'] == 'OK'
-    return false unless @result['results'][0]['address_components'].length > 1
-    true
+    def success?
+      return false unless @result['status'] == 'OK'
+      return false unless @result['results'][0]['address_components'].length > 1
+      true
+    end
   end
 end
