+++ lib/address_geocoder/version.rb
@@ -1,3 +1,3 @@
-module AddressGeocoder
+class AddressGeocoder
   VERSION = '0.0.1'
 end
